# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 回测平台 API
  version: 1.0.0
  description: |
    回测平台提供组合回测、资产分析、蒙特卡洛模拟、组合优化、有效前沿计算、
    战术分配、信号分析、PCA 主成分分析、LETF 滑点分析、目标优化等量化投资工具。

    ## 认证
    - 管理端点（admin、data/manage）需要 `x-api-key` 请求头认证
    - 计算密集型端点（backtest、backtest-optimizer）可选 `x-api-key`，有 Key 时验证身份，无 Key 时放行但记录匿名访问
    - 健康检查和 metrics 端点不需要认证

    ## 速率限制
    - 普通 API：100 次/15 分钟/IP
    - 计算密集型 API（backtest、backtest-optimizer）：10 次/分钟/IP
servers:
  - url: http://localhost:5001/api/v1
    description: 本地开发环境
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 访问令牌，通过 /auth/login 获取
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          oneOf:
            - type: string
            - type: object
              required:
                - code
                - message
              properties:
                code:
                  type: string
                  description: 机器可读错误码
                message:
                  type: string
                  description: 人类可读错误描述

    # ===== 认证鉴权 =====
    # 企业理由：API 契约文档是前后端协作的唯一可靠依据。
    # 认证流程涉及令牌签发、刷新、撤销，任何字段歧义都会导致集成故障，
    # 因此必须将 LoginRequest / TokenResponse / UserInfo 等结构固化到 OpenAPI 规范中，
    # 确保客户端与服务端对令牌生命周期管理达成一致理解。
...
paths:
...
  /data/history:
    get:
      tags: [data]
      summary: 获取历史行情数据
      description: 获取指定资产的历史价格数据
      parameters:
        - name: tickers
          in: query
          required: true
          schema:
            type: string
          description: 逗号分隔的资产代码，如 SPY,VTI
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: 开始日期
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: 结束日期
      responses:
        '200':
          description: 历史数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    description: "{ ticker: { date: price } }"
                    additionalProperties:
                      type: object
                      additionalProperties:
                        type: number
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 数据获取失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /data/search:
    get:
      tags: [data]
      summary: 搜索资产代码
      description: 根据关键词搜索资产代码，支持中英文搜索
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: 搜索关键词
        - name: market
          in: query
          required: false
          schema:
            type: string
          description: 市场筛选（如 A股）
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
        '400':
          description: 缺少 query 参数
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 搜索失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /data/cpi/{country}:
    get:
      tags: [data]
      summary: 获取 CPI 数据
      description: 获取指定国家的消费者价格指数数据
      parameters:
        - name: country
          in: path
          required: true
          schema:
            type: string
            enum: [us, cn]
          description: 国家代码
      responses:
        '200':
          description: CPI 数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                        value:
                          type: number
        '400':
          description: 不支持的国家代码
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: CPI 数据文件不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 数据获取失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===== 数据引擎管理 =====
...